requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
catboost==0.24.1
category-encoders==2.2.2
cloudpickle==1.3.0
colour==0.1.5
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.6
dtreeviz==1.0
fastparquet==0.4.1
graphviz==0.17
gunicorn==20.1.0
importlib-metadata==4.8.1
iniconfig==1.1.1
joblib==0.14.1
kiwisolver==1.3.1
lightgbm==3.0.0
llvmlite==0.36.0
matplotlib==3.3.4
mljar-supervised==0.8.9
numba==0.53.1
numpy==1.19.5
packaging==21.0
pandas==1.1.2
patsy==0.5.2
Pillow==8.4.0
plotly==5.3.1
pluggy==1.0.0
py==1.10.0
pyarrow==5.0.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
shap==0.36.0
six==1.16.0
slicer==0.0.7
sqlparse==0.4.2
statsmodels==0.12.2
tabulate==0.8.7
tenacity==8.0.1
threadpoolctl==3.0.0
thrift==0.15.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
Unipath==1.1
whitenoise==5.3.0
wordcloud==1.7.0
xgboost==1.2.0
zipp==3.6.0
